The jmac httpservletchallenge moved to maven

+++ appserver/tests/application/src/test/java/org/glassfish/main/test/app/security/jmac/http/servlet/challenge/HttpServletChallengeAuthTest.java
@@ -14,18 +14,21 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  */
 
-package org.glassfish.main.test.app.security.jmac.http.servlet.basic;
+package org.glassfish.main.test.app.security.jmac.http.servlet.challenge;
 
 import java.io.File;
 import java.io.InputStream;
 import java.lang.System.Logger;
+import java.net.CookieHandler;
+import java.net.CookieManager;
+import java.net.HttpCookie;
 import java.net.HttpURLConnection;
+import java.net.URLEncoder;
 import java.nio.file.Path;
 import java.util.Base64;
 
 import org.glassfish.main.itest.tools.GlassFishTestEnvironment;
 import org.glassfish.main.itest.tools.asadmin.Asadmin;
-import org.glassfish.main.test.app.security.jmac.http.servlet.basic.HttpServletTestAuthModule;
 import org.jboss.shrinkwrap.api.ShrinkWrap;
 import org.jboss.shrinkwrap.api.exporter.ZipExporter;
 import org.jboss.shrinkwrap.api.spec.JavaArchive;
@@ -37,7 +40,7 @@ import org.junit.jupiter.api.io.TempDir;
 
 import static java.lang.System.Logger.Level.INFO;
 import static java.nio.charset.StandardCharsets.UTF_8;
-import static org.glassfish.main.itest.tools.GlassFishTestEnvironment.USER_PASSWORD;
+import static org.glassfish.main.itest.tools.GlassFishTestEnvironment.createFileUser;
 import static org.glassfish.main.itest.tools.GlassFishTestEnvironment.getDomain1Directory;
 import static org.glassfish.main.itest.tools.GlassFishTestEnvironment.openConnection;
 import static org.glassfish.main.itest.tools.asadmin.AsadminResultMatcher.asadminOK;
@@ -46,14 +49,16 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.stringContainsInOrder;
 
 
-public class HtmlAuthTest {
+public class HttpServletChallengeAuthTest {
+    private static final Logger LOG = System.getLogger(HttpServletChallengeAuthTest.class.getName());
 
-    private static final Logger LOG = System.getLogger(HtmlAuthTest.class.getName());
-
-    private static final String APP_NAME = "security-jmac-httpservlet";
-    private static final String USER_NAME = "shingwai123";
-    private static final String AUTH_MODULE_NAME = "httpServletTestAuthModule";
+    private static final String APP_NAME = "security-jmac-http-servlet-challenge";
+    private static final String AUTH_MODULE_NAME = "httpServletChallengeTestAuthModule";
     private static final String FILE_REALM_NAME = "file123";
+    private static final String USER_NAME = "shingwai";
+    private static final String USER_NAME2 = "shingwai_2";
+    private static final String USER_PASSWORD = "shingwai";
+    private static final String USER_PASSWORD2 = "adminadmin";
 
     private static final Asadmin ASADMIN = GlassFishTestEnvironment.getAsadmin();
 
@@ -66,30 +71,31 @@ public class HtmlAuthTest {
 
     @BeforeAll
     public static void prepareDeployment() {
-        keyFile = getDomain1Directory().resolve(Path.of("config", "keyfile123.txt")).toFile();
+        keyFile = getDomain1Directory().resolve(Path.of("config", "file123.txt")).toFile();
         assertThat(ASADMIN.exec("create-auth-realm", "--classname",
             "com.sun.enterprise.security.auth.realm.file.FileRealm", "--property",
             "file=" + keyFile.getAbsolutePath() + ":jaas-context=fileRealm", "--target", "server", FILE_REALM_NAME),
             asadminOK());
-        assertThat(ASADMIN.exec("create-file-user", "--groups", "mygroup", "--authrealmname", FILE_REALM_NAME,
-            "--target", "server", USER_NAME), asadminOK());
+        createFileUser(FILE_REALM_NAME, USER_NAME, USER_PASSWORD, "mygroup");
+        createFileUser(FILE_REALM_NAME, USER_NAME2, USER_PASSWORD2, "mygroup");
 
-        JavaArchive loginModule = ShrinkWrap.create(JavaArchive.class).addClass(MyHttpServletResponseWrapper.class)
-            .addClass(HttpServletTestAuthModule.class).addClass(MyPrintWriter.class);
+        final JavaArchive loginModule = ShrinkWrap.create(JavaArchive.class)
+            .addClass(HttpServletChallengeTestAuthModule.class);
         LOG.log(INFO, loginModule.toString(true));
+        // FIXME: When you use the same name as in HttpServletBasicAuthTest, the test will fail with HTTP500.
         loginModuleFile = new File(getDomain1Directory().toAbsolutePath().resolve("../../lib").toFile(),
-            "testLoginModule.jar");
+            "testLoginModuleChallenge.jar");
         loginModule.as(ZipExporter.class).exportTo(loginModuleFile, true);
 
         assertThat(ASADMIN.exec("create-message-security-provider",
-            "--classname", HttpServletTestAuthModule.class.getName(),
+            "--classname", HttpServletChallengeTestAuthModule.class.getName(),
             "--layer", "HttpServlet", "--providertype", "server", "--requestauthsource", "sender",
             AUTH_MODULE_NAME), asadminOK());
 
-        WebArchive webArchive = ShrinkWrap.create(WebArchive.class)
-            .addAsWebResource(HtmlAuthTest.class.getPackage(), "index.jsp", "index.jsp")
-            .addAsWebInfResource(HtmlAuthTest.class.getPackage(), "web.xml", "web.xml")
-            .addAsWebInfResource(HtmlAuthTest.class.getPackage(), "glassfish-web.xml", "glassfish-web.xml");
+        final WebArchive webArchive = ShrinkWrap.create(WebArchive.class)
+            .addAsWebResource(HttpServletChallengeAuthTest.class.getPackage(), "index.jsp", "index.jsp")
+            .addAsWebInfResource(HttpServletChallengeAuthTest.class.getPackage(), "web.xml", "web.xml")
+            .addAsWebInfResource(HttpServletChallengeAuthTest.class.getPackage(), "sun-web.xml", "sun-web.xml");
 
         LOG.log(INFO, webArchive.toString(true));
 
@@ -103,6 +109,7 @@ public class HtmlAuthTest {
     @AfterAll
     public static void cleanup() {
         ASADMIN.exec("undeploy", APP_NAME);
+        ASADMIN.exec("delete-file-user", "--authrealmname", FILE_REALM_NAME, "--target", "server", USER_NAME2);
         ASADMIN.exec("delete-file-user", "--authrealmname", FILE_REALM_NAME, "--target", "server", USER_NAME);
         ASADMIN.exec("delete-message-security-provider", "--layer", "HttpServlet", AUTH_MODULE_NAME);
         ASADMIN.exec("delete-auth-realm", FILE_REALM_NAME);
@@ -114,25 +121,40 @@ public class HtmlAuthTest {
 
     @Test
     void test() throws Exception {
-        HttpURLConnection connection = openConnection(8080, "/" + APP_NAME + "/index.jsp");
-        connection.setRequestMethod("GET");
-        String basicAuth = Base64.getEncoder().encodeToString((USER_NAME + ":" + USER_PASSWORD).getBytes(UTF_8));
-        connection.setRequestProperty("Authorization", "Basic " + basicAuth);
-        assertThat(connection.getResponseCode(), equalTo(200));
-        try (InputStream is = connection.getInputStream()) {
-            String text = new String(is.readAllBytes(), UTF_8);
-            assertThat(text, stringContainsInOrder(
-                "Hello World from 196 HttpServlet AuthModule Test!",
-                "Hello, shingwai123 from " + HttpServletTestAuthModule.class.getName(),
-                "PC = security-jmac-httpservlet/security-jmac-httpservlet",
-                "Adjusted count: 230"));
+        final CookieManager cookieManager = new CookieManager();
+        CookieHandler.setDefault(cookieManager);
+        final HttpCookie sessionId;
+        final HttpURLConnection connection = openConnection(8080, "/" + APP_NAME + "/index.jsp");
+        try {
+            connection.setRequestMethod("GET");
+            connection.setDoOutput(true);
+            final String basicAuth = Base64.getEncoder().encodeToString((USER_NAME + ":" + USER_PASSWORD).getBytes(UTF_8));
+            connection.setRequestProperty("Authorization", "Basic " + basicAuth);
+            assertThat(connection.getResponseCode(), equalTo(401));
+            sessionId = cookieManager.getCookieStore().getCookies().stream()
+                .filter(c -> c.getName().equals("JSESSIONID")).findFirst().get();
         } finally {
             connection.disconnect();
         }
+        final HttpURLConnection connection2 = openConnection(8080, "/" + APP_NAME + "/index.jsp");
+        connection2.setRequestProperty("Cookie", "JSESSIONID=" + URLEncoder.encode(sessionId.getValue(), "UTF-8"));
+        connection2.setRequestMethod("GET");
+        final String basicAuth = Base64.getEncoder().encodeToString((USER_NAME + ":" + USER_PASSWORD2).getBytes(UTF_8));
+        connection2.setRequestProperty("Authorization", "Basic " + basicAuth);
+        assertThat(connection2.getResponseCode(), equalTo(200));
+        try (InputStream is = connection2.getInputStream()) {
+            final String text = new String(is.readAllBytes(), UTF_8);
+            assertThat(text, stringContainsInOrder(
+                "Hello World from 196 HttpServletChallenge AuthModule Test!",
+                "Hello, shingwai from " + HttpServletChallengeTestAuthModule.class.getName(),
+                "with authType MC"));
+        } finally {
+            connection2.disconnect();
+        }
     }
 
-    private static void delete(File file) {
-        if (file.exists()) {
+    private static void delete(final File file) {
+        if (file != null && file.exists()) {
             file.delete();
         }
     }
